get image to render
